+++ __tests__/MockClient.ts
@@ -1,8 +1,7 @@
-import {Ed25519PrivateKey} from "../src/Keys";
 import {BaseClient} from "../src/BaseClient";
 import {grpc} from "@improbable-eng/grpc-web";
 import ProtobufMessage = grpc.ProtobufMessage;
-import {CryptoTransferTransaction} from "../exports";
+import {CryptoTransferTransaction, Ed25519PrivateKey} from "../exports";
 
 export const privateKey = Ed25519PrivateKey.fromString('302e020100300506032b657004220420db484b828e64b2d8f12ce3c0a0e93a0b8cce7af1bb8f39c97732394482538e10');
