+++ src/crypto/PublicKey.ts
@@ -1,0 +1,5 @@
+import {Key} from "../generated/BasicTypes_pb";
+
+export interface PublicKey {
+    toProtoKey(): Key;
+}
