+++ src/file/FileCreateTransaction.ts
@@ -6,10 +6,10 @@ import {BaseClient} from "../BaseClient";
 
 import {FileService} from "../generated/FileService_pb_service";
 import {FileCreateTransactionBody} from "../generated/FileCreate_pb";
-import {Ed25519PublicKey} from "../Keys";
 import {KeyList} from "../generated/BasicTypes_pb";
 import {timestampToProto} from "../util";
 import {dateToTimestamp} from "../types/Timestamp";
+import {Ed25519PublicKey} from "../crypto/Ed25519PublicKey";
 
 export class FileCreateTransaction extends TransactionBuilder {
     private readonly body: FileCreateTransactionBody;
