+++ src/file/FileUpdateTransaction.ts
@@ -5,12 +5,12 @@ import {grpc} from "@improbable-eng/grpc-web";
 import {BaseClient} from "../BaseClient";
 
 import {FileService} from "../generated/FileService_pb_service";
-import {PublicKey} from "../Keys";
 import {KeyList} from "../generated/BasicTypes_pb";
 import {FileUpdateTransactionBody} from "../generated/FileUpdate_pb";
 import {timestampToProto} from "../util";
 import {FileIdLike, fileIdToProto} from "../types/FileId";
 import {dateToTimestamp} from "../types/Timestamp";
+import {PublicKey} from "../crypto/PublicKey";
 
 export class FileUpdateTransaction extends TransactionBuilder {
     private readonly body: FileUpdateTransactionBody;
